lab5

+++ src/reservation/reservation.api.controller.ts
@@ -8,6 +8,7 @@ import {
   NotFoundException,
   UsePipes,
   ValidationPipe,
+  BadRequestException,
 } from '@nestjs/common';
 import { ReservationService } from './reservation.service';
 import { CreateReservationDto } from './dto/create-reservation.dto';
@@ -25,7 +26,7 @@ import {
 export class ReservationApiController {
   constructor(private readonly reservationService: ReservationService) {}
 
-  @Post('create/:userId')
+  @Post(':userId')
   @ApiOperation({ summary: 'Создать бронирование для пользователя' })
   @ApiParam({ name: 'userId', type: String, description: 'ID пользователя' })
   @ApiBody({ type: CreateReservationDto })
@@ -34,16 +35,31 @@ export class ReservationApiController {
     @Param('userId') userId: string,
     @Body() dto: CreateReservationDto,
   ) {
-    return this.reservationService.create(userId, dto);
+    if (dto.months <= 0) {
+      throw new BadRequestException(
+        'Количество месяцев должно быть положительным числом',
+      );
+    }
+    try {
+      return await this.reservationService.create(userId, dto);
+    } catch (error) {
+      if (error instanceof NotFoundException) {
+        throw new NotFoundException(error.message);
+      }
+      throw new BadRequestException(
+        'Не удалось создать бронирование, введите корректные данные',
+      );
+    }
   }
 
-  @Get('search/:userId')
+  @Get(':userId')
   @ApiOperation({ summary: 'Получить бронирования пользователя' })
   @ApiParam({ name: 'userId', type: String, description: 'ID пользователя' })
   @ApiResponse({
     status: 200,
     description: 'Список бронирований с привязкой к складам',
   })
+  @ApiResponse({ status: 400, description: 'Бронирование не найдено' })
   async findAll(@Param('userId') userId: string) {
     const reservations = await this.reservationService.findAll(userId);
     if (!reservations || reservations.length === 0) {
@@ -52,12 +68,26 @@ export class ReservationApiController {
     return reservations;
   }
 
-  @Delete('delete/:id')
+  @Delete(':id')
   @ApiOperation({ summary: 'Удалить бронирование' })
   @ApiParam({ name: 'id', type: String })
   @ApiResponse({ status: 200, description: 'Бронирование удалено' })
-  @ApiResponse({ status: 404, description: 'Бронирование не найдено' })
+  @ApiResponse({ status: 400, description: 'Бронирование не найдено' })
   async remove(@Param('id') id: string) {
-    return this.reservationService.remove(Number(id));
+    if (!id || Number(id) <= 0) {
+      throw new BadRequestException(
+        'ID бронирования должен быть положительным числом',
+      );
+    }
+    try {
+      return await this.reservationService.remove(Number(id));
+    } catch (error) {
+      if (error instanceof NotFoundException) {
+        throw new NotFoundException('Бронирование не найдено');
+      }
+      throw new BadRequestException(
+        'Не удалось удалить бронирование, введите правильный id',
+      );
+    }
   }
 }
